Pipeline crates: split giants, retire vestigial CLI, rename ambiguous modules
Three structural cleanups landed together:

- terrain-core: split the two largest modules into focused submodules.
    level.rs    2,008 → level/{format,parse,binary,build,grid_builders}.rs
    terrain.rs  2,227 → terrain/{contour,spatial,idw_grid,lookup_grid,
                                 distance,query}.rs
  All resulting files are <800 lines with a single clear responsibility;
  mod.rs in each subdirectory re-exports the previously-public items so
  consumer imports work unchanged.

- wavemesh-builder: retire the standalone CLI (build-level is the
  canonical entry point), drop two 1-line re-export shims (level.rs,
  terrain.rs), and have consumers import from terrain_core directly.

- Disambiguate the two unrelated marching-* algorithms by file name:
    build-level/src/marching.rs        → marching_squares.rs (raster→contour)
    wavemesh-builder/src/marching.rs   → ray_march.rs (wavefront ray-march)

Also drops build-level/src/simplify.rs (a 7-line type alias that was
left behind after the geometry-helpers consolidation); callers inline
the local Point alias.

No logic edits — purely structural.
