--- pipeline/wavemesh-builder/src/level.rs
@@ -1,1 +1,0 @@
-pub use terrain_core::level::*;
