+++ pipeline/wavemesh-builder/src/ray_march.rs
@@ -9,10 +9,10 @@ use rayon::prelude::*;
 
 use crate::config::MeshBuildConfig;
 use crate::decimate::decimate_track_snapshots;
-use crate::level::TerrainCPUData;
 use crate::physics::{advance_interior_ray, advance_sentinel_ray, RayState};
 use crate::refine::{refine_wavefront, RefineStats};
-use crate::terrain::{ContourLookupGrid, ParsedContour};
+use terrain_core::level::TerrainCPUData;
+use terrain_core::terrain::{ContourLookupGrid, ParsedContour};
 use crate::wavefront::{
     SegmentTrack, SegmentTrackSnapshot, WaveBounds, WaveParams, WavefrontSegment,
 };
