--- pipeline/wavemesh-builder/src/terrain.rs
@@ -1,1 +1,0 @@
-pub use terrain_core::terrain::*;
