+++ pipeline/wavemesh-builder/src/wavefront.rs
@@ -2,7 +2,7 @@
 //! Mirrors marchingTypes.ts, segmentTracks.ts, MeshBuildTypes.ts.
 
 use crate::config::MeshBuildConfig;
-use crate::level::WaveSource;
+use terrain_core::level::WaveSource;
 
 /// Number of f32 values per vertex in the output mesh buffer.
 pub const VERTEX_FLOATS: usize = 6;
